Added "Merciless" AI. Can now play 1v1 with CPU
Low-level AI implemented. Game still cannot be "won" because it has no
win condition, but pieces can be kinged and jumped. The low level AI
evaluates the CPU's pieces starting from the lowest board index, and
takes the first jump it finds. If no jump is available, it takes the
first regular move it can find.