Solved bug where whole board is filled up with black pieces on move 2
Emptypiece was being overwritten to have a "black" color instead of
none, and so whenever it was copied to empty a square, instead of
emptying it, a black piece would be placed there. Solved by creating
objects for each use of emptypiece, rather than using one global object.
Will refactor next. Lots of useless code remains in the js page