Version 0.1
Finally, a working demo. Although in need of a lot of refactoring,
checkers0.1 is finally able to complete a full game with a human
opponent while also using animated moves and the MiniMax algorithm.
Minimax runs at depth of 2, because the code as it stands is very
inefficient, and anything higher takes 10+ seconds per move. Over the
next few days, will be redesigning and refactoring to make it
application faster, and the code simpler.